Credited myself in credits tab
Also, reminder: this version of STARK only works with Team Fortress 2, I
did this because selectedGame is broken, and tf2 is the only source game
I play so yeah. You can change the game by pressing CTRL + SHIFT + F in
Visual Studio and search for "tf\" in the entire solution, then replace
the results with the source game of your choice.