Added ability to replace words or sentences with other words or sentences
There is a new file called "replace.txt" and you can use it to replace
words or sentences.
For example: if you put "test=cookies" in the file, it will change the
word "test" into "cookies" in tts messages.
You can also do this: "this is a test=I like pizza", this will change
"this is a test" into "I like pizza".
"this is a test=I like pizza" (enter) "cookies=donuts", this will change
"this is a test" into "I like pizza", and it will also turn "cookies"
into "donuts".
Also, it is NOT capital sensitive.