Added new tab and new commands
-Added new tab "Commands"
In this tab you can choose what commands should not be available to
non-whitelisted users.
-Added command ".skip"
This command will skip the current message in the queue. This command is
only available to whitelisted users by default, you can change that in
the commands tab.
-Added command ".clear"
This command will clear all messages in the queue. This command is only
available to whitelisted users by default, you can change that in the
commands tab.