Added even more commands
-Added command ".block"
This command will block someone.
For example: if you type ".block annoying person", it will add "annoying
person" to blocked_users.txt.
This means that you no longer have to put the player name in
blocked_users.txt, you can now do it from in-game.
This command can only be used by whitelisted users by default, you can
change that in the "Commands" tab.
-Added command ".blockword"
This command allows you to block a word or sentence.
It works pretty much the same as ".block" but it adds the word or
sentence to blocked_words.txt.
This command can only be used by whitelisted users by default, you can
change that in the "Commands" tab.