Added YouTube Video support!
-YouTube Video playback is now possible by using the command:
".playvideo <Video URL or Video ID>" (it downloads the audio of the
videos using the YoutubeExplode library and converts them to mp3 with
the MediaToolkit library)
-added CheckBox in Audio tab called "Allow .play and .playvideo while a
song is playing." If it's checked, you can switch songs, basically. If
it's not checked, the commands ".play" and ".playvideo" will not work
while a song is playing.