Audio queue, YouTube search, and other new stuff
- .play and .playvideo will now add songs to the audio queue, so that
means there is a queue now.
- The .skip command skips the current song and plays the next song in
the audio queue.
- The .ttsskip command skips the current TTS message in the TTS queue.
- The .clear command clears the audio queue except for the currently
playing song.
- The .ttsclear command clears the TTS queue.
- New tab: Audio queue
- Now you can search and play the first video using .playvideo <tags>.
(new library: YoutubeSearch)
- YoutubeExplode library has been updated because older versions didn't
work anymore (it also added some other libraries: AngleSharp and
Newtonsoft.Json)
- The default sample rate has been changed to 44100 with 2 channels
(much better quality). You can change the sample rate in the file
"audioformat.txt".
- In the Audio tab there is a new button called "Add to queue" which
adds the selected item in the audio file list to the audio queue.
- In the Audio tab the "Allow .play and .playvideo while a
song is playing." has been removed because it's pretty useless now that
we have an audio queue.
- Cleaned up code in CommandReader.cs because it was a mess (I made a
function that checks if a player executing a command actually is allowed
to do so, and I used that function for all commands instead of
copy-pasting all the code).

Note: for some reason .play <tag> doesn't work anymore (I don't know if
it's just me). I tried debugging and stuff but couldn't figure out why
it doesn't work anymore, but .play is kinda deprecated anyway since
there is .playvideo.

Note 2: There seems to be a bug with the audio queue: when you pause and
unpause a song a few times and wait till it's over, the song doesn't go
away. Just press Skip when this happens, or just don't spam Pause.